# gdal raster tile: speed-up generation of (max zoom) tiles in PNG format