… autotest/utilities/data/test_gdal_raster_tile_paeth.jpg
… autotest/utilities/data/test_gdal_raster_tile_paeth.jpg.aux.xml