By using an optimized code path, one can speed by 36% in some cases tile generation, when using libdeflate. If using zlib only, the gain is more modest, but still there.

e.g using  https://nz-imagery.s3-ap-southeast-2.amazonaws.com/otago/otago_2025_0.2m/rgb/2193/CA14_5000_0304.tiff

Before:
```
$ time GDAL_RASTER_TILE_USE_PNG_OPTIM=NO gdal raster tile CA14_5000_0304.tiff CA14_5000_0304_tiled --progress
0...10...20...30...40...50...60...70...80...90...100 - done in 00:00:12.

real    0m13,234s
user    2m14,696s
sys     0m2,220s

Size of output directory: 373.356 MB
```

After:
```
$ time gdal raster tile CA14_5000_0304.tiff CA14_5000_0304_tiled --progress
0...10...20...30...40...50...60...70...80...90...100 - done in 00:00:08.

real    0m8,352s
user    1m26,253s
sys     0m1,929s

Size of output directory: 359.384 MB
```