@jcphill 2025-07-16
The tile problem is fixed and the runtime is down to 11:10 now, so 17% overall improvement.