@jcphill 2025-07-16
The max zoom tile png files are 3% or so larger, though.