@jratike80 2025-07-16
> The max zoom tile png files are 5% or so larger, though.

What if you add `--co Z_LEVEL=7`? It should compress better, but perhaps it eats the improvement in speed.