@jcphill 2025-07-16
> > The max zoom tile png files are 5% or so larger, though.
> 
> What if you add `--co Z_LEVEL=7`? It should compress better, but perhaps it eats the improvement in speed.

Yes, --co ZLEVEL=7 is 4% smaller and 15% slower, although I'm seeing inconsistent timings today.