@rouault 2025-07-29
@jcphill Could you retry now that I've implement the PAETH PNG filter ?