@jcphill 2025-07-31
> > Runtime up to 12:49 (basically same as without this PR),
> 
> with libdeflate ?

Yes.