@rouault 2025-08-05
I've implemented a SSE2 version of the PAETH filter that recovers most of the performance losses due to the introduction of PAETH.