@jcphill 2025-08-08
Now it's even slower, up to 14:01. These are AMD Milan processors. Repeating runs to be sure.