@jcphill 2025-08-08
Reran back-to-back in same job. July 3 version took 11:52, latest version took 10:37. Still 2.8% larger than July 3 version.