@rouault 2025-08-08
> Was the output supposed to be identical?

yes, last commit is supposed to be a pure speed optimization, so there's a bug somewhere, likely in the cost evaluation of PAETH vs average filter