@jcphill 2025-08-08
Size is back down where it was. Have a good weekend!